refactor: move url_validation to core to honor layer dependency rule
pipeline/download.py and worker/stage_tasks.py imported web.url_validation, an upward dependency that contradicts the README's documented layer rule (core <- pipeline <- worker <- web). The module is regex-only with no web-framework coupling, so it belongs in core; relocating it lets pipeline/worker/web all import downward into core. No behaviour change.
