+++ src/whisper_ui/web/routes/upload.py
@@ -14,6 +14,19 @@ from markupsafe import escape
 from whisper_ui.core.constants import MAX_BATCH_SIZE
 from whisper_ui.core.languages import DEFAULT_WHISPER_MODEL, LANGUAGE_CHOICES, WHISPER_MODELS
 from whisper_ui.core.models import Job, JobStatus
+from whisper_ui.core.url_validation import (
+    GoogleDriveURLError,
+    TwitterURLError,
+    UnsupportedPlaylistTypeError,
+    YouTubeURLError,
+    is_google_drive_url,
+    is_twitter_url,
+    is_youtube_playlist_url,
+    validate_google_drive_url,
+    validate_twitter_url,
+    validate_youtube_playlist_url,
+    validate_youtube_url,
+)
 from whisper_ui.pipeline.audio_probe import get_audio_duration_seconds
 from whisper_ui.pipeline.preprocess import SUPPORTED_EXTENSIONS
 from whisper_ui.ui import labels as ui_labels
@@ -26,19 +39,6 @@ from whisper_ui.web.playlist import (
     PlaylistTooLargeError,
     PlaylistUnavailableError,
     expand_playlist,
-)
-from whisper_ui.web.url_validation import (
-    GoogleDriveURLError,
-    TwitterURLError,
-    UnsupportedPlaylistTypeError,
-    YouTubeURLError,
-    is_google_drive_url,
-    is_twitter_url,
-    is_youtube_playlist_url,
-    validate_google_drive_url,
-    validate_twitter_url,
-    validate_youtube_playlist_url,
-    validate_youtube_url,
 )
 from whisper_ui.web.validation import clamp_num_speakers
 from whisper_ui.worker.pipeline_dispatcher import enqueue_pipeline
